[#29] refactor : interfaces 모듈 ktlint 적용
