+++ coupon-platform-interfaces/src/main/kotlin/coupon_platform/interfaces/couponcode/CouponCodeController.kt
@@ -1,8 +1,8 @@
-package coupon_platform.interfaces.coupon_code
+package coupon_platform.interfaces.couponcode
 
 import coupon_platform.application.couponcode.CouponCodeFacade
 import coupon_platform.interfaces.common.response.BaseResponse
-import coupon_platform.interfaces.coupon_code.dto.CouponCodeCreateRequest
+import coupon_platform.interfaces.couponcode.dto.CouponCodeCreateRequest
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
