+++ coupon-platform-interfaces/src/main/kotlin/coupon_platform/interfaces/couponstats/CouponStatsController.kt
@@ -1,10 +1,10 @@
-package coupon_platform.interfaces.coupon_stats
+package coupon_platform.interfaces.couponstats
 
 import coupon_platform.application.couponstats.CouponStatsFacade
 import coupon_platform.application.couponstats.CouponStatsFacadeV2
 import coupon_platform.domain.couponstats.dto.CouponStatsInfo
 import coupon_platform.interfaces.common.response.BaseResponse
-import coupon_platform.interfaces.coupon_stats.dto.CouponStatsResponse
+import coupon_platform.interfaces.couponstats.dto.CouponStatsResponse
 import org.springframework.web.bind.annotation.*
 import java.time.LocalDate
