+++ coupon-platform-interfaces/src/test/kotlin/coupon_platform/interfaces/support/ExtensionFunctions.kt
